@@ -2,8 +2,7 @@ class PasswordResetsController < ApplicationController
   before_filter :load_user_using_perishable_token, :only => [:edit, :update]
   before_filter :require_no_user, :only => [:edit, :update]
   skip_before_filter :login_check
-  skip_before_filter CASClient::Frameworks::Rails::Filter, :if => Proc.new{|s| s.using_CAS? && @appconfig.login_options.include?('CAS')}
-
+  skip_before_filter RubyCAS::Filter, :if => Proc.new{|s| s.using_CAS? && @appconfig.login_options.include?('CAS')}
 
   def new
     render
