@@ -33,7 +33,7 @@ class ShiftsController < ApplicationController
     @visible_loc_groups = current_user.user_config.view_loc_groups
     @selected_loc_groups = @visible_loc_groups.collect{|l| l.id}
     @visible_locations = current_user.user_config.view_loc_groups.collect{|l| l.locations}.flatten
-    
+
 
     # @calendars = @department.calendars.active
     # @shifts = []
@@ -226,7 +226,7 @@ class ShiftsController < ApplicationController
       end
     end
   end
-  
+
   def email_group
     default_email_group_settings
     @loc_groups = current_department.loc_groups.active
@@ -257,7 +257,7 @@ class ShiftsController < ApplicationController
     else
       @start_time ||= department_day_start_time
     end
-    
+
     if params[:email_group]["end_time(1i)"]
       @end_time ||= DateTime.new(params[:email_group][:"end_time(1i)"].to_i,params[:email_group][:"end_time(2i)"].to_i,params[:email_group][:"end_time(3i)"].to_i,params[:email_group][:"end_time(4i)"].to_i,params[:email_group][:"end_time(5i)"].to_i)
     else
