@@ -2,9 +2,9 @@ module UsersHelper
 
   def deactivate_link(user = @user)
     active = user.is_active?(current_department)
-    link_to_remote (active ? "Deactivate" : "Restore"), {:url => toggle_user_path(user),
+    link_to (active ? "Deactivate" : "Restore"), {:url => toggle_user_path(user),
       :success => "$('#user#{user.id}').toggleClass('disabled');$('#user#{user.id} a.toggle_link').text(($('#user#{user.id}').hasClass('disabled') ? 'Restore' : 'Deactivate'))"},
-      :href => toggle_user_path(user), :class => 'toggle_link'
+      :href => toggle_user_path(user), :remote => true, :class => 'toggle_link'
   end
 
   def determine_class(user)
