@@ -10,8 +10,8 @@ class Calendar < ActiveRecord::Base
 
   validates_uniqueness_of :name, :scope => :department_id
 
-  named_scope :active, lambda {{ :conditions => {:active => true}}}
-  named_scope :public, lambda {{ :conditions => {:public => true}}}
+  scope :active, lambda {{ :conditions => {:active => true}}}
+  scope :public, lambda {{ :conditions => {:public => true}}}
 
   def self.active_in(department, start_date = Time.now, end_date = Time.now)
     active = Calendar.find(:all, :conditions => ["department_id = ? and start_date <= ? and end_date >= ? and active is true", department.id, start_date.utc, end_date.utc])
