@@ -20,10 +20,10 @@ class LocGroup < ActiveRecord::Base
   has_and_belongs_to_many :stickies,      :join_table => :loc_groups_notices, :association_foreign_key => :notice_id
   has_and_belongs_to_many :notices
   
-  named_scope :active, :conditions => {:active => true}
+  scope :active, :conditions => {:active => true}
   
-  before_validation_on_create :create_permissions
-  before_validation_on_update :update_permissions
+  before_validation(:on => :create) {:create_permissions}
+  before_validation(:on => :update) {:update_permissions}
 
   validates_presence_of :department
 
