@@ -1,16 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-# Your secret key for verifying cookie session data integrity.
-# If you change this key, all old sessions will become invalid!
-# Make sure the secret is at least 30 characters and all random,
-# no regular words or you'll be exposed to dictionary attacks.
-ActionController::Base.session = {
-  :key         => '_test_session',
-  :secret      => '152799450ecf87bc47bb4de6b1b7a31bcfbbd99fc945ecb2a49659aeb045e3249a815ef164afd81480eac9be41c907b1aeed7d4ac7738b96bacafd02b48f007c'
-}
+Shifts::Application.config.session_store :cookie_store, :key => '_shifts_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
-# (create the session table with "rake db:sessions:create")
-# ActionController::Base.session_store = :active_record_store
-
+# (create the session table with "rails generate session_migration")
+# Shifts::Application.config.session_store :active_record_store
